requirements.txt
Flask
flask-cors
google-generativeai
python-dotenv
pytesseract
Pillow